+++ www/templates/film.tpl
@@ -28,12 +28,12 @@
 				<?= $film['description'] ?>
 			</p>
 		</div>
-
-		<div class="film__footer">
-			<a href="<?= HOST ?>edit.php?id=<?= $film['id'] ?>" class="btn btn--edit">Редактировать</a>
-			<a href="<?= HOST ?>delete.php?id=<?= $film['id'] ?>" class="btn btn--delete">Удалить</a>
-		</div>
-
+		<?php if (isAdmin()): ?>
+			<div class="film__footer">
+				<a href="<?= HOST ?>edit.php?id=<?= $film['id'] ?>" class="btn btn--edit">Редактировать</a>
+				<a href="<?= HOST ?>delete.php?id=<?= $film['id'] ?>" class="btn btn--delete">Удалить</a>
+			</div>
+		<?php endif; ?>
 	</div>
 	<!-- // Desc -->
 </article>
